Implement several NaturalTransformations
Of course it was possible to satisfy the original question. The transformation
from IO to Try is obvious once it's written, and the lack of a `performIO`
method on IO even more vexing.

Stylistically, I like the way the code reads when the transformations are all
in gerund form. E.g. performingIO().andThen(throwingExceptions()) just flows so
naturally and is so descriptive of the operation.

It was fairly straightforward to implement a number of different standard
NaturalTransformations that operate on the monad types in lambda. A couple of
patterns emerged, but I think there are probably edge cases that are also
possible. In general, it's possible to change the argument monad of a MonadT.
Where there are related Monad and MonadT types - e.g. Either and EitherT - it
is generally posible to transform the Monad into the MonadT given a Pure for
the argument monad.

Some monads and monad transformers also produce a different monad with the same
carrier when they're run, such as SafeT. For these is was possible to implement
"runner" transformations. I can readily imagine use cases for these at the
point in an application where the monad stack is finally dispensed. This made
me wonder if there's another higher-order type of function that could provide a
framework for cases where there is a transformation from one functor to
andother and also one carrier type to another. I'm thinking specifically of
running monad transformers like EitherT and StateT, but there are probably lots
of use cases where a bespoke higher-order transformation could be really useful.

Finally, after writing a bunch of these, I had a couple of ideas for helper
functions that panned out nicely and I needed to write some new hamcrest
matchers for testing. To accomodate these, I've reorganized sly into submodules
to create test-scoped sub libraries and also start to divide up functionality
into related groups in order to enforce non-cyclic dependency and also let
downstream code pick and choose what to include.

Here is the complete list of new NaturalTransformations:

- PerformingIO: IO -> Try
- MappingStreamT: StreamT<M, ?> -> StreamT<N, ?>
- MappingMaybeT: MaybeT<M, ?> -> MaybeT<N, ?>
- MappingEitherT: EitherT<M, L, ?> -> EitherT<N, L, ?>
- MappingStateT: StateT<S, M, ?> -> StateT<S, N, ?>
- MappingIterateT: IterateT<M, ?> -> IterateT<N, ?>
- MappingReaderT: ReaderT<R, M, ?> -> ReaderT<R, N, ?>
- MappingSafeT: SafeT<M, ?> -> SafeT<N, ?>
- MappingIdentityT: IdentityT<M, ?> -> IdentityT<N, ?>
- RunningReaderT: ReaderT<R, M, ?> -> M
- RunningSafeT: SafeT<M, ?> -> M
- RunningIdentityT: IdentityT<M, ?> -> M
- TransformingMaybe: Maybe<?> -> MaybeT<M, ?>
- TransformingEither: Either<L, ?> -> EitherT<M, L, ?>
- TransformingState: State<S, ?> -> StateT<S, M, ?>
- ThrowingExceptions: Try -> Identity

Here is the complete list of new Matchers and other test-scoped code:

- TryMatcher: Matcher of Try
- StreamTMatcher: Matcher of StreamT
- IterableMatcher: Matcher of Iterable
- JustMatcher: Matcher of Maybe
- StateMatcher: Matcher of State
- DescriptionOf, DescribeMismatch and SelfDescribing: helpers for extracting
  hamcrest object descriptions
- IndentingDescription: a decorator for a hamcrest Description that allows a
  matcher to indent the descriptions of delegate matchers
- YoloIO: a transformation that kept coming up in testing

Other additions:

- Stringy: because sometimes you just want a CharSequence to also be
  Iterable<Character> and to be able to easily convert between String,
  CharSequence, and Iterable<Character>
- FromPure: create a transformation from Identity to a functor given a Pure
- ToFn: a helper to treat a transformation as Fn1<FA, GA>