Implement RunningEitherT
Create a new interface HyperFn that expresses a transformation of the form F a -> G b with separate parameterization for all four types. This is sufficiently generic to capture the operation of running an instance of EitherT, in which the Functor transforms from EitherT<M, L, ?> to MonadRec<?, M> and the carrier from R to Either<L, R> and the two transformations are inextricably linked to one another.