Joining effects across transformers
I don't know if this is a useful concept or not, but there are Monad/MonadT pairs that are the same monadic effect. Here are some HyperFns that join the effect of a transformer with its corresponding non-transformer, when it occurs in the argument monad.

I couldn't quite figure out how to test a couple of them. Ok, let's be real, JoiningIdentity is kind of silly and not worth testing.