sly-app (#7)
Create a functional framework for building an application from separate services.
Services can reference other services as dependencies using handles. The built 
application is resolved in the form of a `Runner` for short-lived application or
a `Starter` for long-lived applications to create an application function. This
is then either run or started using a static method of `App`.